+++ test/main.js
@@ -3,13 +3,17 @@ import { serialize, parse } from 'error-serializer'
 
 import { SIMPLE_ERROR_OBJECT } from './helpers/main.js'
 
-test('Normalize invalid types when parsing', (t) => {
+test('Normalize invalid types when parsing with "normalize: true"', (t) => {
   const message = 'test'
-  t.is(parse(message).message, message)
+  t.is(parse(message, { normalize: true }).message, message)
 })
 
 test('Normalize core properties when parsing', (t) => {
-  t.is(typeof parse({ ...SIMPLE_ERROR_OBJECT, stack: 0 }).stack, 'string')
+  t.is(
+    typeof parse({ ...SIMPLE_ERROR_OBJECT, stack: 0 }, { normalize: true })
+      .stack,
+    'string',
+  )
 })
 
 test('Remove unsafe non-core properties when serializing', (t) => {